pass/requirements.txt
aioredis==2.0.1
amqp==5.1.1
anyio==3.7.0
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
fastapi==0.96.0
fernet==1.0.1
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
kombu==5.3.1
priority==2.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyaes==1.6.1
pycparser==2.21
pydantic==1.10.8
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
sqlalchemy-cockroachdb==2.0.1
SQLAlchemy-serializer==1.4.1
starlette==0.27.0
toml==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
vine==5.0.0
wallet-py3k==0.0.4
wcwidth==0.2.6
wsproto==1.2.0
